--- data/src/main/kotlin/photos/network/data/user/network/entity/ApiResponse.kt
@@ -1,6 +1,0 @@
-package photos.network.data.user.network.entity
-
-@kotlinx.serialization.Serializable
-data class ApiResponse(
-    val message: String,
-)
