--- data/src/main/kotlin/photos/network/data/user/network/entity/VerifyClientIdResponse.kt
@@ -1,11 +1,0 @@
-package photos.network.data.user.network.entity
-
-import kotlinx.serialization.SerialName
-
-@kotlinx.serialization.Serializable
-data class VerifyClientIdResponse(
-    val error: String,
-
-    @SerialName("error_description")
-    val errorDescription: String
-)
