+++ data/src/main/kotlin/photos/network/data/user/network/model/ApiResponse.kt
@@ -1,0 +1,8 @@
+package photos.network.data.user.network.model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ApiResponse(
+    val message: String,
+)
