+++ data/src/main/kotlin/photos/network/data/user/network/model/TokenInfo.kt
@@ -13,15 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package photos.network.data.user.network.entity
+package photos.network.data.user.network.model
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class TokenResponse(
+data class TokenInfo(
     @SerialName("access_token") val accessToken: String,
-    @SerialName("token_type") val tokenType: String? = null,
-    @SerialName("expires_in") val expiresIn: Long,
-    @SerialName("refresh_token") val refreshToken: String
+    @SerialName("expires_in") val expiresIn: Int,
+    @SerialName("refresh_token") val refreshToken: String,
+    @SerialName("token_type") val tokenType: String,
 )
